Refactor user.py into utils/user.py
