--- hazmapper_plugin/pb_tool.cfg
@@ -1,80 +1,0 @@
-#/***************************************************************************
-# HazmapperPlugin
-#
-# Configuration file for plugin builder tool (pb_tool)
-# Generated by Plugin Builder: http://g-sherman.github.io/Qgis-Plugin-Builder/
-#                            -------------------
-#       begin               : 2025-05-02
-#       copyright           : (C) 2025 by DesignSafe-CI
-#       email               : designsafe-ci@tacc.utexas.edu
-# ***************************************************************************/
-#
-#/***************************************************************************
-# *                                                                         *
-# *   This program is free software; you can redistribute it and/or modify  *
-# *   it under the terms of the GNU General Public License as published by  *
-# *   the Free Software Foundation; either version 2 of the License, or     *
-# *   (at your option) any later version.                                   *
-# *                                                                         *
-# ***************************************************************************/
-# 
-# 
-# You can install pb_tool using:
-#  pip install http://geoapt.net/files/pb_tool.zip
-# 
-# Consider doing your development (and install of pb_tool) in a virtualenv.
-# 
-# For details on setting up and using pb_tool, see:
-#  http://g-sherman.github.io/plugin_build_tool/
-#
-# Issues and pull requests here:
-# https://github.com/g-sherman/plugin_build_tool:
-# 
-# Sane defaults for your plugin generated by the Plugin Builder are
-# already set below.
-# 
-# As you add Python source files and UI files to your plugin, add
-# them to the appropriate [files] section below.
-
-[plugin]
-# Name of the plugin. This is the name of the directory that will
-# be created in .qgis2/python/plugins
-name: hazmapper_plugin
-
-# Full path to where you want your plugin directory copied. If empty,
-# the QGIS default path will be used. Don't include the plugin name in
-# the path.
-plugin_path:
-
-[files]
-# Python  files that should be deployed with the plugin
-python_files: __init__.py hazmapper_plugin.py hazmapper_plugin_dockwidget.py
-
-# The main dialog file that is loaded (not compiled)
-main_dialog: hazmapper_plugin_dockwidget_base.ui
-
-# Other ui files for dialogs you create (these will be compiled)
-compiled_ui_files: 
-
-# Resource file(s) that will be compiled
-resource_files: resources.qrc
-
-# Other files required for the plugin
-extras: metadata.txt icon.png
-
-# Other directories to be deployed with the plugin.
-# These must be subdirectories under the plugin directory
-extra_dirs:
-
-# ISO code(s) for any locales (translations), separated by spaces.
-# Corresponding .ts files must exist in the i18n directory
-locales:
-
-[help]
-# the built help directory that should be deployed with the plugin
-dir: help/build/html
-# the name of the directory to target in the deployed plugin 
-target: help
-
-
-
